fix: atualizada todas as referencias de i32 para f64 para evitar problema no calculo de area

+++ src/geometria.rs
@@ -1,12 +1,12 @@
 pub mod formas_geometricas {
     pub struct Triangulo {
-        a: i32,
-        b: i32,
-        c: i32
+        a: f64,
+        b: f64,
+        c: f64
     }
 
     impl Triangulo {
-        pub fn new(a: i32, b: i32, c: i32) -> Triangulo {
+        pub fn new(a: f64, b: f64, c: f64) -> Triangulo {
             if ((a + b) <= c) || ((a + c) <= b) || ((b + c) <= a) {
                 panic!("Se a soma entre os dois lados é igual ou menor ao terceiro, esse triângulo não pode existir.")
             }
@@ -14,7 +14,7 @@ pub mod formas_geometricas {
             Triangulo { a, b, c }
         }
 
-        pub fn equals(&self, a: i32, b: i32, c: i32) -> bool {
+        pub fn equals(&self, a: f64, b: f64, c: f64) -> bool {
             self.a == a && self.b == b && self.c == c
         }
 
@@ -32,7 +32,7 @@ pub mod formas_geometricas {
             self.a == self.b && self.b == self.c
         }
 
-        pub fn perimetro(&self) -> i32 {
+        pub fn perimetro(&self) -> f64 {
             self.a + self.b + self.c
         }
 
@@ -41,9 +41,9 @@ pub mod formas_geometricas {
             // p = (a + b + c) / 2
             // A = √p(p-a)(p-b)(p-c)
 
-            let p = (self.a + self.b + self.c) / 2;
+            let p = (self.a + self.b + self.c) / 2_f64;
             
-            ((p * (p - self.a) * (p - self.b) * (p - self.c)) as f64).sqrt()
+            (p * (p - self.a) * (p - self.b) * (p - self.c)).sqrt()
         }
     }
 }
